Vincent's endpoint changes; moved refresh code and other things to helper
