+++ dockstore-webservice/src/main/java/io/dockstore/webservice/resources/DockerRepoResource.java
@@ -17,12 +17,10 @@
 package io.dockstore.webservice.resources;
 
 import com.codahale.metrics.annotation.Timed;
-import com.esotericsoftware.yamlbeans.YamlReader;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Optional;
 import com.google.gson.Gson;
 import io.dockstore.webservice.Helper;
-import io.dockstore.webservice.api.UserRequest;
 import io.dockstore.webservice.core.Container;
 import io.dockstore.webservice.core.Tag;
 import io.dockstore.webservice.core.Token;
@@ -39,15 +37,11 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Base64;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
@@ -94,31 +88,6 @@ public class DockerRepoResource {
 
     private final List<String> namespaces = new ArrayList<>();
 
-    private static class RepoList {
-
-        private List<Container> repositories;
-
-        public void setRepositories(List<Container> repositories) {
-            this.repositories = repositories;
-        }
-
-        public List<Container> getRepositories() {
-            return this.repositories;
-        }
-    }
-
-    private static class Collab {
-        private String content;
-
-        public void setContent(String content) {
-            this.content = content;
-        }
-
-        public String getContent() {
-            return this.content;
-        }
-    }
-
     public DockerRepoResource(ObjectMapper mapper, HttpClient client, UserDAO userDAO, TokenDAO tokenDAO, ContainerDAO containerDAO,
             TagDAO tagDAO) {
         this.objectMapper = mapper;
@@ -141,226 +110,27 @@ public class DockerRepoResource {
         namespaces.add("pancancer");
     }
 
-    @GET
-    @Path("/user/{userId}")
-    @Timed
-    @UnitOfWork
-    @ApiOperation(value = "List repos owned by the logged-in user", notes = "Lists all registered and unregistered containers owned by the user", response = Container.class, responseContainer = "List")
-    public List<Container> userContainers(@ApiParam(hidden = true) @Auth Token token,
-            @ApiParam(value = "User ID", required = true) @PathParam("userId") Long userId) {
-        User user = userDAO.findById(token.getUserId());
-        Helper.checkUser(user, userId);
-
-        List<Container> ownedContainers = containerDAO.findByUserId(userId);
-        return ownedContainers;
-    }
-
     @PUT
     @Path("/refresh")
     @Timed
     @UnitOfWork
-    @ApiOperation(value = "Refresh repos owned by the logged-in user", notes = "Updates some metadata", response = Container.class, responseContainer = "List")
-    @SuppressWarnings("checkstyle:methodlength")
-    public List<Container> refresh(@ApiParam(hidden = true) @Auth Token authToken,
-            @ApiParam(value = "UserRequest to refresh the list of repos for a user", required = true) UserRequest request) {
-        long userId = request.getId();
-
+    @ApiOperation(value = "Refresh all repos", notes = "Updates some metadata. ADMIN ONLY", response = Container.class, responseContainer = "List")
+    // @SuppressWarnings("checkstyle:methodlength")
+    public List<Container> refreshAll(@ApiParam(hidden = true) @Auth Token authToken) {
         User authUser = userDAO.findById(authToken.getUserId());
-        Helper.checkUser(authUser, userId);
+        Helper.checkUser(authUser);
 
-        List<Container> currentRepos = containerDAO.findByUserId(userId);
-        List<Container> allRepos = new ArrayList<>(0);
-        List<Token> tokens = tokenDAO.findByUserId(userId);
-        Map<String, ArrayList> tagMap = new HashMap<>();
-
-        SimpleDateFormat formatter = new SimpleDateFormat("EEE, d MMM yyyy HH:mm:ss Z");
-
-        List<String> namespaceList = new ArrayList<>();
-
-        Token quayToken = null;
-        Token gitToken = null;
-
-        // Get user's quay and git tokens
-        for (Token token : tokens) {
-            if (token.getTokenSource().equals(TokenType.QUAY_IO.toString())) {
-                quayToken = token;
-            }
-            if (token.getTokenSource().equals(TokenType.GITHUB_COM.toString())) {
-                gitToken = token;
+        List<Container> containers = new ArrayList<>();
+        List<User> users = userDAO.findAll();
+        for (User user : users) {
+            try {
+                containers.addAll(Helper.refresh(user.getId(), client, objectMapper, namespaces, LOG, userDAO, containerDAO, tokenDAO,
+                        tagDAO));
+            } catch (WebApplicationException ex) {
+                LOG.info("Failed to refresh user " + user.getId());
             }
         }
-
-        if (gitToken == null || quayToken == null) {
-            LOG.info("GIT or QUAY token not found!");
-            throw new WebApplicationException(HttpStatus.SC_CONFLICT);
-        }
-
-        namespaceList.add(quayToken.getUsername());
-        namespaceList.addAll(namespaces);
-
-        GitHubClient githubClient = new GitHubClient();
-        githubClient.setOAuth2Token(gitToken.getContent());
-        try {
-            UserService uService = new UserService(githubClient);
-            OrganizationService oService = new OrganizationService(githubClient);
-            RepositoryService service = new RepositoryService(githubClient);
-            ContentsService cService = new ContentsService(githubClient);
-            org.eclipse.egit.github.core.User user = uService.getUser();
-
-            // for (String namespace : namespaces) {
-            for (String namespace : namespaceList) {
-                String url = TARGET_URL + "repository?namespace=" + namespace;
-                Optional<String> asString = ResourceUtilities.asString(url, quayToken.getContent(), client);
-
-                if (asString.isPresent()) {
-                    RepoList repos;
-                    try {
-                        repos = objectMapper.readValue(asString.get(), RepoList.class);
-                        LOG.info("RESOURCE CALL: " + url);
-
-                        List<Container> containers = repos.getRepositories();
-
-                        for (Container c : containers) {
-                            String repo = c.getNamespace() + "/" + c.getName();
-                            String path = quayToken.getTokenSource() + "/" + repo;
-
-                            // Get the list of builds from the container.
-                            // Builds contain information such as the Git URL and tags
-                            String urlBuilds = TARGET_URL + "repository/" + repo + "/build/";
-                            Optional<String> asStringBuilds = ResourceUtilities.asString(urlBuilds, quayToken.getContent(), client);
-
-                            String gitURL = "";
-
-                            if (asStringBuilds.isPresent()) {
-                                String json = asStringBuilds.get();
-                                LOG.info("RESOURCE CALL: " + urlBuilds);
-
-                                // parse json using Gson to get the git url of repository and the list of tags
-                                Gson gson = new Gson();
-                                Map<String, ArrayList> map = new HashMap<>();
-                                map = (Map<String, ArrayList>) gson.fromJson(json, map.getClass());
-                                ArrayList builds = map.get("builds");
-
-                                if (!builds.isEmpty()) {
-                                    Map<String, Map<String, String>> map2 = new HashMap<>();
-                                    map2 = (Map<String, Map<String, String>>) builds.get(0);
-
-                                    gitURL = map2.get("trigger_metadata").get("git_url");
-
-                                    Map<String, String> map3 = (Map<String, String>) builds.get(0);
-                                    String lastBuild = (String) map3.get("started");
-                                    LOG.info("LAST BUILD: " + lastBuild);
-
-                                    Date date = null;
-                                    try {
-                                        date = formatter.parse(lastBuild);
-                                        c.setLastBuild(date);
-                                    } catch (ParseException ex) {
-                                        LOG.info("Build date did not match format 'EEE, d MMM yyyy HH:mm:ss Z'");
-                                    }
-
-                                    tagMap.put(path, (ArrayList<String>) map2.get("tags"));
-                                }
-                            }
-
-                            c.setRegistry(quayToken.getTokenSource());
-                            c.setGitUrl(gitURL);
-
-                            List<Repository> gitRepos = new ArrayList<>(0);
-                            gitRepos.addAll(service.getRepositories(user.getLogin()));
-
-                            for (org.eclipse.egit.github.core.User org : oService.getOrganizations()) {
-                                gitRepos.addAll(service.getRepositories(org.getLogin()));
-                            }
-
-                            for (Repository repository : gitRepos) {
-                                LOG.info(repository.getSshUrl());
-                                if (repository.getSshUrl().equals(c.getGitUrl())) {
-                                    try {
-                                        List<RepositoryContents> contents = cService.getContents(repository, "collab.cwl");
-                                        if (!(contents == null || contents.isEmpty())) {
-                                            c.setHasCollab(true);
-
-                                            String encoded = contents.get(0).getContent().replace("\n", "");
-                                            byte[] decode = Base64.getDecoder().decode(encoded);
-                                            String content = new String(decode, StandardCharsets.UTF_8);
-
-                                            // parse the collab.cwl file to get description and author
-                                            YamlReader reader = new YamlReader(content);
-                                            Object object = reader.read();
-                                            Map map = (Map) object;
-                                            String description = (String) map.get("description");
-                                            map = (Map) map.get("dct:creator");
-                                            String author = (String) map.get("foaf:name");
-
-                                            c.setDescription(description);
-                                            c.setAuthor(author);
-
-                                            LOG.info("Repo: " + repository.getName() + " has collab.cwl");
-                                        }
-                                    } catch (IOException ex) {
-                                        LOG.info("Repo: " + repository.getName() + " has no collab.cwl");
-                                    }
-                                }
-                            }
-                        }
-
-                        allRepos.addAll(containers);
-                    } catch (IOException ex) {
-                        LOG.info("Exception: " + ex);
-                    }
-                }
-            }
-        } catch (IOException ex) {
-            LOG.info("Token ignored due to IOException: " + gitToken.getId() + " " + ex);
-        }
-
-        Date time = new Date();
-
-        for (Container newContainer : allRepos) {
-            boolean exists = false;
-
-            for (Container oldContainer : currentRepos) {
-                if (newContainer.getName().equals(oldContainer.getName())
-                        && newContainer.getNamespace().equals(oldContainer.getNamespace())
-                        && newContainer.getRegistry().equals(oldContainer.getRegistry())) {
-                    exists = true;
-
-                    oldContainer.update(newContainer);
-
-                    break;
-                }
-            }
-
-            if (!exists) {
-                newContainer.setUserId(userId);
-                String path = newContainer.getRegistry() + "/" + newContainer.getNamespace() + "/" + newContainer.getName();
-                newContainer.setPath(path);
-
-                currentRepos.add(newContainer);
-            }
-        }
-
-        for (Container container : currentRepos) {
-            container.setLastUpdated(time);
-            containerDAO.create(container);
-
-            container.getTags().clear();
-
-            ArrayList<String> tags = tagMap.get(container.getPath());
-            if (tags != null) {
-                for (String tag : tags) {
-                    LOG.info("Creating tag: " + tag);
-                    Tag newTag = new Tag();
-                    newTag.setVersion(tag);
-                    long tagId = tagDAO.create(newTag);
-                    newTag = tagDAO.findById(tagId);
-                    container.addTag(newTag);
-                }
-            }
-        }
-
-        return currentRepos;
+        return containers;
     }
 
     @GET
@@ -383,8 +153,22 @@ public class DockerRepoResource {
     public Container getContainer(@ApiParam(hidden = true) @Auth Token authToken,
             @ApiParam(value = "Container ID", required = true) @PathParam("containerId") Long containerId) {
         Container c = containerDAO.findById(containerId);
+        Helper.checkContainer(c);
+
         User user = userDAO.findById(authToken.getUserId());
         Helper.checkUser(user, c.getUserId());
+
+        return c;
+    }
+
+    @GET
+    @Timed
+    @UnitOfWork
+    @Path("/registered/{containerId}")
+    @ApiOperation(value = "Get a registered container", notes = "NO authentication", response = Container.class)
+    public Container getRegisteredContainer(@ApiParam(value = "Container ID", required = true) @PathParam("containerId") Long containerId) {
+        Container c = containerDAO.findRegisteredById(containerId);
+        Helper.checkContainer(c);
 
         return c;
     }
@@ -392,18 +176,15 @@ public class DockerRepoResource {
     @POST
     @Timed
     @UnitOfWork
-    @Path("/register")
+    @Path("/{containerId}/register")
     @ApiOperation(value = "Register a container", notes = "Register a container (public or private). Assumes that user is using quay.io and github. Include quay.io in path if using quay.io", response = Container.class)
-    public Container register(@ApiParam(hidden = true) @Auth Token authToken, @QueryParam("repository") String path,
-            @QueryParam("enduser_id") Long userId) {
+    public Container register(@ApiParam(hidden = true) @Auth Token authToken,
+            @ApiParam(value = "Container id to delete", required = true) @PathParam("containerId") Long containerId) {
+        Container c = containerDAO.findById(containerId);
+        Helper.checkContainer(c);
+
         User user = userDAO.findById(authToken.getUserId());
-        Helper.checkUser(user, userId);
-
-        Container c = containerDAO.findByPath(path);
-
-        if (c == null || c.getUserId() != userId) {
-            throw new WebApplicationException(HttpStatus.SC_BAD_REQUEST);
-        }
+        Helper.checkUser(user, c.getUserId());
 
         if (c.getHasCollab() && !c.getGitUrl().isEmpty()) {
             c.setIsRegistered(true);
@@ -415,14 +196,16 @@ public class DockerRepoResource {
         }
     }
 
-    @DELETE
+    @POST
     @UnitOfWork
-    @Path("/unregister/{containerId}")
-    @ApiOperation(value = "Deletes a container", hidden = true)
+    @Path("/{containerId}/unregister")
+    @ApiOperation(value = "Unregister a container", response = Container.class)
     public Container unregister(@ApiParam(hidden = true) @Auth Token authToken,
             @ApiParam(value = "Container id to delete", required = true) @PathParam("containerId") Long containerId) {
-        User user = userDAO.findById(authToken.getUserId());
         Container c = containerDAO.findById(containerId);
+        Helper.checkContainer(c);
+
+        User user = userDAO.findById(authToken.getUserId());
         Helper.checkUser(user, c.getUserId());
 
         c.setIsRegistered(false);
@@ -434,21 +217,7 @@ public class DockerRepoResource {
     @GET
     @Timed
     @UnitOfWork
-    @Path("/allRegistered/{userId}")
-    @ApiOperation(value = "List all registered containers from a user", notes = "Get user's registered containers only", response = Container.class, responseContainer = "List")
-    public List<Container> userRegisteredContainers(@ApiParam(hidden = true) @Auth Token authToken,
-            @ApiParam(value = "User ID", required = true) @PathParam("userId") Long userId) {
-        User user = userDAO.findById(authToken.getUserId());
-        Helper.checkUser(user, userId);
-
-        List<Container> repositories = containerDAO.findRegisteredByUserId(userId);
-        return repositories;
-    }
-
-    @GET
-    @Timed
-    @UnitOfWork
-    @Path("allRegistered")
+    @Path("registered")
     @ApiOperation(value = "List all registered containers. This would be a minimal resource that would need to be implemented "
             + "by a GA4GH reference server", tags = { "GA4GH", "containers" }, notes = "NO authentication", response = Container.class, responseContainer = "List")
     public List<Container> allRegisteredContainers() {
@@ -459,11 +228,16 @@ public class DockerRepoResource {
     @GET
     @Timed
     @UnitOfWork
-    @Path("registered")
-    @ApiOperation(value = "Get a registered container", notes = "Lists info of container. Enter full path (include quay.io in path). NO authentication", response = Container.class)
-    public Container getRegisteredContainer(@QueryParam("repository") String repo) {
-        Container repository = containerDAO.findRegisteredByPath(repo);
-        return repository;
+    @Path("/path/{path}")
+    @ApiOperation(value = "Get a registered container by path", notes = "Lists info of container. Enter full path (include quay.io in path).", response = Container.class)
+    public Container getRegisteredContainer(@ApiParam(hidden = true) @Auth Token authToken, @QueryParam("repository") String path) {
+        Container container = containerDAO.findRegisteredByPath(path);
+        Helper.checkContainer(container);
+
+        User user = userDAO.findById(authToken.getUserId());
+        Helper.checkUser(user, container.getUserId());
+
+        return container;
     }
 
     @PUT
@@ -551,6 +325,7 @@ public class DockerRepoResource {
     @ApiOperation(value = "List the tags for a registered container", response = Tag.class, responseContainer = "List", hidden = true)
     public List<Tag> tags(@ApiParam(hidden = true) @Auth Token authToken, @QueryParam("containerId") long containerId) {
         Container repository = containerDAO.findById(containerId);
+        Helper.checkContainer(repository);
 
         User user = userDAO.findById(authToken.getUserId());
         Helper.checkUser(user, repository.getUserId());
@@ -563,45 +338,71 @@ public class DockerRepoResource {
     @GET
     @Timed
     @UnitOfWork
-    @Path("/collab")
-    @ApiOperation(value = "Get the corresponding collab.cwl file on Github", notes = "Enter full path of container (add quay.io if using quay.io)", response = Collab.class)
-    public Collab collab(@ApiParam(hidden = true) @Auth Token authToken, @QueryParam("repository") String repository) {
-        Container container = containerDAO.findByPath(repository);
+    @Path("/{containerId}/collab")
+    @ApiOperation(value = "Get the corresponding collab.cwl file on Github", notes = "Enter full path of container (add quay.io if using quay.io)", response = Helper.Collab.class)
+    public Helper.Collab collab(@ApiParam(hidden = true) @Auth Token authToken,
+            @ApiParam(value = "Container id to delete", required = true) @PathParam("containerId") Long containerId) {
+        Container container = containerDAO.findById(containerId);
+        Helper.checkContainer(container);
 
         User authUser = userDAO.findById(authToken.getUserId());
-        // Helper.checkUser(authUser, container.getUserId());
+        Helper.checkUser(authUser, container.getUserId());
 
         boolean hasGithub = false;
 
-        Collab collab = new Collab();
+        Helper.Collab collab = new Helper.Collab();
 
-        // StringBuilder builder = new StringBuilder();
-        if (container != null) {
-            Helper.checkUser(authUser, container.getUserId()); // null check first
+        if (!container.getIsRegistered()) {
+            LOG.info("Container " + container.getPath() + " is not registered!");
+            throw new WebApplicationException(HttpStatus.SC_BAD_REQUEST);
+        }
 
-            List<Token> tokens = tokenDAO.findByUserId(container.getUserId());
+        List<Token> tokens = tokenDAO.findByUserId(container.getUserId());
 
-            for (Token token : tokens) {
-                if (token.getTokenSource().equals(TokenType.GITHUB_COM.toString())) {
-                    hasGithub = true;
-                    GitHubClient githubClient = new GitHubClient();
-                    githubClient.setOAuth2Token(token.getContent());
-                    try {
-                        UserService uService = new UserService(githubClient);
-                        OrganizationService oService = new OrganizationService(githubClient);
-                        RepositoryService service = new RepositoryService(githubClient);
-                        ContentsService cService = new ContentsService(githubClient);
-                        org.eclipse.egit.github.core.User user = uService.getUser();
-                        // builder.append("Token: ").append(token.getId()).append(" is ").append(user.getName()).append(" login is ")
-                        // .append(user.getLogin()).append("\n");
+        for (Token token : tokens) {
+            if (token.getTokenSource().equals(TokenType.GITHUB_COM.toString())) {
+                hasGithub = true;
+                GitHubClient githubClient = new GitHubClient();
+                githubClient.setOAuth2Token(token.getContent());
+                try {
+                    UserService uService = new UserService(githubClient);
+                    OrganizationService oService = new OrganizationService(githubClient);
+                    RepositoryService service = new RepositoryService(githubClient);
+                    ContentsService cService = new ContentsService(githubClient);
+                    org.eclipse.egit.github.core.User user = uService.getUser();
+                    // builder.append("Token: ").append(token.getId()).append(" is ").append(user.getName()).append(" login is ")
+                    // .append(user.getLogin()).append("\n");
 
-                        // look through user's own repositories
-                        for (Repository repo : service.getRepositories(user.getLogin())) {
-                            // LOG.info(repo.getGitUrl());
-                            // LOG.info(repo.getHtmlUrl());
-                            LOG.info(repo.getSshUrl()); // ssh url example: git@github.com:userspace/name.git
-                            // LOG.info(repo.getUrl());
-                            // LOG.info(container.getGitUrl());
+                    // look through user's own repositories
+                    for (Repository repo : service.getRepositories(user.getLogin())) {
+                        // LOG.info(repo.getGitUrl());
+                        // LOG.info(repo.getHtmlUrl());
+                        LOG.info(repo.getSshUrl()); // ssh url example: git@github.com:userspace/name.git
+                        // LOG.info(repo.getUrl());
+                        // LOG.info(container.getGitUrl());
+                        if (repo.getSshUrl().equals(container.getGitUrl())) {
+                            try {
+                                List<RepositoryContents> contents = cService.getContents(repo, "collab.cwl");
+                                // odd, throws exceptions if file does not exist
+                                if (!(contents == null || contents.isEmpty())) {
+                                    String encoded = contents.get(0).getContent().replace("\n", "");
+                                    byte[] decode = Base64.getDecoder().decode(encoded);
+                                    String content = new String(decode, StandardCharsets.UTF_8);
+                                    // builder.append(content);
+                                    collab.setContent(content);
+                                }
+                            } catch (IOException ex) {
+                                // builder.append("Repo: ").append(repo.getName()).append(" has no collab.cwl \n");
+                                LOG.info("Repo: " + repo.getName() + " has no collab.cwl");
+                            }
+                        }
+                    }
+
+                    // looks through all repos from different organizations user is in
+                    List<org.eclipse.egit.github.core.User> organizations = oService.getOrganizations();
+                    for (org.eclipse.egit.github.core.User org : organizations) {
+                        for (Repository repo : service.getRepositories(org.getLogin())) {
+                            LOG.info(repo.getSshUrl());
                             if (repo.getSshUrl().equals(container.getGitUrl())) {
                                 try {
                                     List<RepositoryContents> contents = cService.getContents(repo, "collab.cwl");
@@ -619,46 +420,17 @@ public class DockerRepoResource {
                                 }
                             }
                         }
-
-                        // looks through all repos from different organizations user is in
-                        List<org.eclipse.egit.github.core.User> organizations = oService.getOrganizations();
-                        for (org.eclipse.egit.github.core.User org : organizations) {
-                            for (Repository repo : service.getRepositories(org.getLogin())) {
-                                LOG.info(repo.getSshUrl());
-                                if (repo.getSshUrl().equals(container.getGitUrl())) {
-                                    try {
-                                        List<RepositoryContents> contents = cService.getContents(repo, "collab.cwl");
-                                        // odd, throws exceptions if file does not exist
-                                        if (!(contents == null || contents.isEmpty())) {
-                                            String encoded = contents.get(0).getContent().replace("\n", "");
-                                            byte[] decode = Base64.getDecoder().decode(encoded);
-                                            String content = new String(decode, StandardCharsets.UTF_8);
-                                            // builder.append(content);
-                                            collab.setContent(content);
-                                        }
-                                    } catch (IOException ex) {
-                                        // builder.append("Repo: ").append(repo.getName()).append(" has no collab.cwl \n");
-                                        LOG.info("Repo: " + repo.getName() + " has no collab.cwl");
-                                    }
-                                }
-                            }
-                        }
-
-                    } catch (IOException ex) {
-                        // builder.append("Token ignored due to IOException: ").append(token.getId()).append("\n");
-                        LOG.info("Token ignored due to IOException: " + token.getId());
                     }
+
+                } catch (IOException ex) {
+                    // builder.append("Token ignored due to IOException: ").append(token.getId()).append("\n");
+                    LOG.info("Token ignored due to IOException: " + token.getId());
                 }
             }
-            if (!hasGithub) {
-                // builder.append("Github is not setup");
-                LOG.info("Github is not setup");
-                throw new WebApplicationException(HttpStatus.SC_BAD_REQUEST);
-            }
-
-        } else {
-            // builder.append(repository).append(" is not registered");
-            LOG.info(repository + " is not registered");
+        }
+        if (!hasGithub) {
+            // builder.append("Github is not setup");
+            LOG.info("Github is not setup");
             throw new WebApplicationException(HttpStatus.SC_BAD_REQUEST);
         }
